refactor: align repos to new base
